Add ordinal suffixes, some edits, toys.
Edited snippets.cpp to include more #ifndef/define/endif sets for later debug
purposes. (This allows eg: #ifdef EPSILON, do something with epsilon). This was
done as a matter of course as I added an ordinal suffixes helper function. (Two,
actually: int and float.) These automatically generate "st" for 1, "nd" for 2,
etc. The float version basically says, if the float doesn't need to be a float
(ie: equals an int), cast to an int and use that version of the function;
otherwise, the suffix is "th." This approximates actual English, in which
fractional ordinals get a "th" suffix. Technically, there is no right way (or,
at least, no way sounds right) to suffix an ordinal with a fractional part
represented as digits to the right of the decimal point, but a computer can't
hem and haw and say, "You know what I mean," so this is my best approximation.

Also, apparently I somehow had a bunch of toys not added to source control. Or,
possibly, altered without being updated in source control. This update fixes
that.