Add triangle.cpp to toys.
All triangle.cpp does is print the nth triangle number and its corresponding
triangle. Because I want the triangle to look pretty, n is limited by the width
of the terminal. If the user does not provide n, the default is 9. (The 9th
triangle number is 45, btw.) I know I describe my toys as "one-hour-like
projects," but this one took either five minutes or four hours, depending on
whether or not you count how long it took me to write the code to generate
appropriate ordinal suffixes (pulled out and added to snippets before I moved
triangle to my toys folder). Most of that time was spent rat-holing into the
dangerous wilds of preprocessor macros, in hopes of making the suffix function
slightly more portable, but given it relies on runtime code, that really didn't
work out too well. I ended up with a class version (which is what I posted to my
Facebook to entertain my friends), but immediately after realized a single
function returning a const was better in every way.

Oh well; I think a Java programmer would have appreciated my class one, at
least. Except that I actually used a struct, so 1) it's not called a class (the
horror!) and 2), all members are public by default (oh teh noes!). Sorry if
you're a Java programmer; I'm really more making fun of Java advocates. Program
in whatever way is comfortable for you; just don't try to advocate objectively
less-efficient (for both the computer and the programmer) paradigms to people
who already have a preferred paradigm. Okay, that's not really necessary for a
commit message, but it's not like this is for a professional project or
anything.